refactored router patch
